unit-tests: move primMetadata edit restrictions test to a better place
